+++ packages/clerk-js/src/core/resources/CommerceSubscription.ts
@@ -1,6 +1,6 @@
 import type {
   CancelSubscriptionParams,
-  CommerceMoney,
+  CommerceMoneyAmount,
   CommerceSubscriptionItemJSON,
   CommerceSubscriptionItemResource,
   CommerceSubscriptionJSON,
@@ -12,7 +12,7 @@ import type {
 
 import { unixEpochToDate } from '@/utils/date';
 
-import { commerceMoneyFromJSON } from '../../utils';
+import { commerceMoneyAmountFromJSON } from '../../utils';
 import { BaseResource, CommercePlan, DeletedObject } from './internal';
 
 export class CommerceSubscription extends BaseResource implements CommerceSubscriptionResource {
@@ -23,7 +23,7 @@ export class CommerceSubscription extends BaseResource implements CommerceSubscr
   pastDueAt!: Date | null;
   updatedAt!: Date | null;
   nextPayment: {
-    amount: CommerceMoney;
+    amount: CommerceMoneyAmount;
     date: Date;
   } | null = null;
   subscriptionItems!: CommerceSubscriptionItemResource[];
@@ -47,12 +47,12 @@ export class CommerceSubscription extends BaseResource implements CommerceSubscr
     this.pastDueAt = data.past_due_at ? unixEpochToDate(data.past_due_at) : null;
     this.nextPayment = data.next_payment
       ? {
-          amount: commerceMoneyFromJSON(data.next_payment.amount),
+          amount: commerceMoneyAmountFromJSON(data.next_payment.amount),
           date: unixEpochToDate(data.next_payment.date),
         }
       : null;
     this.subscriptionItems = (data.subscription_items || []).map(item => new CommerceSubscriptionItem(item));
-    this.eligibleForFreeTrial = data.eligible_for_free_trial;
+    this.eligibleForFreeTrial = this.withDefault(data.eligible_for_free_trial, false);
     return this;
   }
 }
@@ -69,9 +69,9 @@ export class CommerceSubscriptionItem extends BaseResource implements CommerceSu
   canceledAt!: Date | null;
   pastDueAt!: Date | null;
   //TODO(@COMMERCE): Why can this be undefined ?
-  amount?: CommerceMoney;
+  amount?: CommerceMoneyAmount;
   credit?: {
-    amount: CommerceMoney;
+    amount: CommerceMoneyAmount;
   };
   isFreeTrial!: boolean;
 
@@ -98,8 +98,9 @@ export class CommerceSubscriptionItem extends BaseResource implements CommerceSu
     this.periodEnd = data.period_end ? unixEpochToDate(data.period_end) : null;
     this.canceledAt = data.canceled_at ? unixEpochToDate(data.canceled_at) : null;
 
-    this.amount = data.amount ? commerceMoneyFromJSON(data.amount) : undefined;
-    this.credit = data.credit && data.credit.amount ? { amount: commerceMoneyFromJSON(data.credit.amount) } : undefined;
+    this.amount = data.amount ? commerceMoneyAmountFromJSON(data.amount) : undefined;
+    this.credit =
+      data.credit && data.credit.amount ? { amount: commerceMoneyAmountFromJSON(data.credit.amount) } : undefined;
 
     this.isFreeTrial = this.withDefault(data.is_free_trial, false);
     return this;