+++ packages/clerk-js/src/ui/components/Checkout/CheckoutForm.tsx
@@ -1,5 +1,5 @@
 import { __experimental_useCheckout as useCheckout, useOrganization } from '@clerk/shared/react';
-import type { CommerceMoney, CommercePaymentSourceResource, ConfirmCheckoutParams } from '@clerk/types';
+import type { CommerceMoneyAmount, CommercePaymentSourceResource, ConfirmCheckoutParams } from '@clerk/types';
 import { useMemo, useState } from 'react';
 
 import { Card } from '@/ui/elements/Card';
@@ -36,6 +36,8 @@ export const CheckoutForm = withCardStateProvider(() => {
   const showPastDue = !!totals.pastDue?.amount && totals.pastDue.amount > 0;
   const showDowngradeInfo = !isImmediatePlanChange;
 
+  const fee = planPeriod === 'month' ? plan.fee : plan.annualMonthlyFee;
+
   return (
     <Drawer.Body>
       <Box
@@ -60,7 +62,7 @@ export const CheckoutForm = withCardStateProvider(() => {
             />
             <LineItems.Description
               prefix={planPeriod === 'annual' ? 'x12' : undefined}
-              text={`${plan.currencySymbol}${planPeriod === 'month' ? plan.amountFormatted : plan.annualMonthlyAmountFormatted}`}
+              text={`${fee.currencySymbol}${fee.amountFormatted}`}
               suffix={localizationKeys('commerce.checkout.perMonth')}
             />
           </LineItems.Group>
@@ -92,7 +94,7 @@ export const CheckoutForm = withCardStateProvider(() => {
             </LineItems.Group>
           )}
 
-          {freeTrialEndsAt && plan.freeTrialDays && (
+          {!!freeTrialEndsAt && !!plan.freeTrialDays && (
             <LineItems.Group variant='tertiary'>
               <LineItems.Title
                 title={localizationKeys('commerce.checkout.totalDueAfterTrial', {
@@ -341,7 +343,7 @@ const ExistingPaymentSourceForm = withCardStateProvider(
     totalDueNow,
     paymentSources,
   }: {
-    totalDueNow: CommerceMoney;
+    totalDueNow: CommerceMoneyAmount;
     paymentSources: CommercePaymentSourceResource[];
   }) => {
     const submitLabel = useSubmitLabel();