+++ packages/types/src/json.ts
@@ -71,13 +71,14 @@ export interface ImageJSON {
 }
 
 export interface EnvironmentJSON extends ClerkResourceJSON {
-  auth_config: AuthConfigJSON;
   api_keys_settings: APIKeysSettingsJSON;
+  auth_config: AuthConfigJSON;
+  client_debug_mode?: boolean;
   commerce_settings: CommerceSettingsJSON;
   display_config: DisplayConfigJSON;
-  user_settings: UserSettingsJSON;
-  organization_settings: OrganizationSettingsJSON;
   maintenance_mode: boolean;
+  organization_settings: OrganizationSettingsJSON;
+  user_settings: UserSettingsJSON;
 }
 
 export interface ClientJSON extends ClerkResourceJSON {
@@ -632,6 +633,9 @@ export interface CommercePlanJSON extends ClerkResourceJSON {
   object: 'commerce_plan';
   id: string;
   name: string;
+  fee: CommerceMoneyAmountJSON;
+  annual_fee: CommerceMoneyAmountJSON;
+  annual_monthly_fee: CommerceMoneyAmountJSON;
   amount: number;
   amount_formatted: string;
   annual_amount: number;
@@ -747,7 +751,7 @@ export interface CommerceStatementGroupJSON extends ClerkResourceJSON {
 export interface CommercePaymentJSON extends ClerkResourceJSON {
   object: 'commerce_payment';
   id: string;
-  amount: CommerceMoneyJSON;
+  amount: CommerceMoneyAmountJSON;
   paid_at?: number;
   failed_at?: number;
   updated_at: number;
@@ -769,9 +773,9 @@ export interface CommercePaymentJSON extends ClerkResourceJSON {
 export interface CommerceSubscriptionItemJSON extends ClerkResourceJSON {
   object: 'commerce_subscription_item';
   id: string;
-  amount?: CommerceMoneyJSON;
+  amount?: CommerceMoneyAmountJSON;
   credit?: {
-    amount: CommerceMoneyJSON;
+    amount: CommerceMoneyAmountJSON;
   };
   payment_source_id: string;
   plan: CommercePlanJSON;
@@ -804,7 +808,7 @@ export interface CommerceSubscriptionJSON extends ClerkResourceJSON {
    * Describes the details for the next payment cycle. It is `undefined` for subscription items that are cancelled or on the free plan.
    */
   next_payment?: {
-    amount: CommerceMoneyJSON;
+    amount: CommerceMoneyAmountJSON;
     date: number;
   };
   /**
@@ -827,7 +831,7 @@ export interface CommerceSubscriptionJSON extends ClerkResourceJSON {
  * <ClerkProvider clerkJsVersion="x.x.x" />
  * ```
  */
-export interface CommerceMoneyJSON {
+export interface CommerceMoneyAmountJSON {
   amount: number;
   amount_formatted: string;
   currency: string;
@@ -843,11 +847,11 @@ export interface CommerceMoneyJSON {
  * ```
  */
 export interface CommerceCheckoutTotalsJSON {
-  grand_total: CommerceMoneyJSON;
-  subtotal: CommerceMoneyJSON;
-  tax_total: CommerceMoneyJSON;
-  total_due_now: CommerceMoneyJSON;
-  credit: CommerceMoneyJSON;
+  grand_total: CommerceMoneyAmountJSON;
+  subtotal: CommerceMoneyAmountJSON;
+  tax_total: CommerceMoneyAmountJSON;
+  total_due_now: CommerceMoneyAmountJSON;
+  credit: CommerceMoneyAmountJSON;
 }
 
 /**